Reorg config
The route and middleware specific to the database recorder has been moved into the datatabase recorder.
