+++ composer.json
@@ -12,7 +12,7 @@
     "homepage": "https://github.com/ericpridham/flow",
     "keywords": ["Laravel", "Flow"],
     "require": {
-        "illuminate/support": "~5|~6|~7",
+        "illuminate/support": "~5|~6|~7|~8|~9",
         "ext-json": "*"
     },
     "require-dev": {
