+++ readme.md
@@ -1,11 +1,6 @@
 # Flow
 
-[![Latest Version on Packagist][ico-version]][link-packagist]
-[![Total Downloads][ico-downloads]][link-downloads]
-[![Build Status][ico-travis]][link-travis]
-[![StyleCI][ico-styleci]][link-styleci]
-
-This is where your description should go. Take a look at [contributing.md](contributing.md) to see a to do list.
+An auto-instrumentation package library for Laravel.
 
 ## Installation
 
@@ -15,43 +10,13 @@ Via Composer
 $ composer require ericpridham/flow
 ```
 
-## Usage
-
-## Change log
-
-Please see the [changelog](changelog.md) for more information on what has changed recently.
-
-## Testing
+## Setup
 
 ``` bash
-$ composer test
+$ php artisan migrate
+$ php artisan flow:install
 ```
 
-## Contributing
+## Usage
 
-Please see [contributing.md](contributing.md) for details and a todolist.
-
-## Security
-
-If you discover any security related issues, please email author email instead of using the issue tracker.
-
-## Credits
-
-- [author name][link-author]
-- [All Contributors][link-contributors]
-
-## License
-
-license. Please see the [license file](license.md) for more information.
-
-[ico-version]: https://img.shields.io/packagist/v/ericpridham/flow.svg?style=flat-square
-[ico-downloads]: https://img.shields.io/packagist/dt/ericpridham/flow.svg?style=flat-square
-[ico-travis]: https://img.shields.io/travis/ericpridham/flow/master.svg?style=flat-square
-[ico-styleci]: https://styleci.io/repos/12345678/shield
-
-[link-packagist]: https://packagist.org/packages/ericpridham/flow
-[link-downloads]: https://packagist.org/packages/ericpridham/flow
-[link-travis]: https://travis-ci.org/ericpridham/flow
-[link-styleci]: https://styleci.io/repos/12345678
-[link-author]: https://github.com/ericpridham
-[link-contributors]: ../../contributors
+Go to `https://your-site/flow`
