+++ src/Flow.php
@@ -56,6 +56,7 @@ class Flow
                 $recorderInstance = $recorder;
             }
             $recorderInstance->init($initParams);
+            $recorderInstance->loadRoutes();
             $this->recorders->push($recorderInstance);
         });
     }
