--- src/Http/routes.php
@@ -1,4 +1,0 @@
-<?php
-
-Route::get('/', 'FlowController@index');
-Route::get('/events', 'FlowController@events');
